Move error handling out of ApplicationController
